+++ js/helpers/utils.ts
@@ -3,7 +3,7 @@ import type { BriefLocation, LocationAndQuery } from "@/types/utils.types";
 /**
  * Checks whether the given value is a string
  */
-export const isString = (value: unknown): boolean => {
+export const isString = (value: unknown): value is string => {
     return ((typeof value === 'string') || (value instanceof String))
 }